pyproject.toml
[project]
name = "local_stock_price_database"
version = "0.1.0"
description = "Local stock price database with Alpaca ingestion, DuckDB/Parquet storage, FastAPI interface."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "duckdb>=1.1.0",
  "pandas>=2.2",
  "pyarrow>=17.0",
  "pydantic-settings>=2.3",
  "orjson>=3.10",
]

[project.optional-dependencies]
dev = ["black>=24.4", "ruff>=0.6"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
